Move api handlers into api package and their own files.
